Refactor the client side into modules
I didn't like how the "manipulating the DOM" code was mixed in with the
business logic, so I split them into separate components.

This refactoring also made it possible to add some validation to the
form (since the "read from the form" code is now all in the same place).