Fix two bugs in mutate-route
First, it would error out if lengthStoreThreshold was a floating-point
number. (Unlike some other parameters, lengthStoreThreshold is allowed
to be a float.)

Second, it would occasionally error out because it was trying to do a
batch request with no items. (We fixed this problem by checking if there
are any items to write.)